Add a new version of the search() method and add a sample main.
The new version create up to 2^n processes to analize the
file to find the keyword given. It's a bit slow because of the
read_line() function. I also add a sample main file. Inside
it there is a particolar loop that prevent the process to exit
until all the children has died.

There is no way to stop the program to generate to much child,
so, be careful to use it.